Split overage DTOs
