+++ tenant-platform/tenant-config/src/main/java/com/lms/tenant/config/TenantConnectionInterceptor.java
@@ -1,5 +1,6 @@
 package com.lms.tenant.config;
 
+import com.common.context.TenantContext;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
@@ -33,7 +34,7 @@ public class TenantConnectionInterceptor extends DelegatingDataSource {
         if (connection == null) {
             return;
         }
-        var tenantId = TenantContext.getTenantId();
+        var tenantId = TenantContext.get();
         if (tenantId.isPresent()) {
             try (Statement statement = connection.createStatement()) {
                 statement.execute("SET LOCAL app.current_tenant = '" + tenantId.get() + "'");
