+++ tenant-platform/tenant-core/src/main/java/com/lms/tenant/core/dto/OverageResponse.java
@@ -1,0 +1,17 @@
+package com.lms.tenant.core.dto;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record OverageResponse(
+        UUID overageId,
+        UUID tenantId,
+        String featureKey,
+        long quantity,
+        long unitPriceMinor,
+        String currency,
+        Instant occurredAt,
+        Instant periodStart,
+        Instant periodEnd,
+        String status) {
+}
