+++ tenant-platform/tenant-core/src/main/java/com/lms/tenant/core/dto/RecordOverageRequest.java
@@ -6,7 +6,6 @@ import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 import java.time.Instant;
 import java.util.Map;
-import java.util.UUID;
 
 public record RecordOverageRequest(
         @NotBlank String featureKey,
@@ -18,17 +17,4 @@ public record RecordOverageRequest(
         @NotNull Instant periodEnd,
         String idempotencyKey,
         Map<String, Object> metadata) {
-}
-
-public record OverageResponse(
-        UUID overageId,
-        UUID tenantId,
-        String featureKey,
-        long quantity,
-        long unitPriceMinor,
-        String currency,
-        Instant occurredAt,
-        Instant periodStart,
-        Instant periodEnd,
-        String status) {
 }
